2.1 Attribute Binding and Event Handling
All right. Let's review another example in this video. We will tackle attribute binding and then also a little bit of a vent handling so it should be fun. Let's get going.

Maybe we have a button here and I'll set the label to click me.

All right. Let's have a look in the browser.

And sure enough we see our button in the top left corner now real quick. Why don't we put it in the center of the page? So I'm going to add a few Styles here and we'll say about this HTML and body should have a height of a hundred percent.

And then on the body itself, I'm going to set a display of grid and then I will call Place items Center. So this is an easy way in 2022 to Center an element on the page.

Come back. Give it a refresh and there we go, perfect horizontal and vertical centering.

Okay.

Next why don't we add a couple classes to control the styling of this button and I'm going to use basic utilities. For example text red will have a color of red and then we'll do another one for how about text green.

Okay, great. So now of course as you know, I can add a class of text green and if we switch back, we now have a green button.