2.2 Okay, but now what if we want to apply these classes dynamically and maybe when you click the button we change the classes even further. How exactly do we do that? All right. Have a look.
The first step is to add a single source of Truth. So just like we did before I'll add a data method that returns an object. And for now, we'll call it button classes.

And I'll set this to text green.

Now in the last video we learned about the mustache braces. So you might think that we could do this but actually no we cannot use them with an attributes.

And in fact, if we try come back give it a refresh and no the text is not green. And in fact, if we take a look at this sure enough, we see the string button classes rather than the the evaluated expression.